backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
django-filter==23.1
djangorestframework-simplejwt==5.2.2
Pillow==9.5.0
python-dotenv==1.0.0
gunicorn==20.1.0
psycopg2-binary==2.9.6
drf-spectacular==0.26.2
django-extensions==3.2.3
celery==5.3.1
redis==4.5.5
